requirements.txt
pandas
python-dotenv
streamlit==1.34.0
streamlit-card
streamlit-extras
numpy

